refactor: segment code
