--- file-tracker/file_tracker/utils.py
@@ -1,3 +1,0 @@
-def pad(list, new_size, padding=b'0'):
-    assert new_size >= len(list)
-    return list + [padding] * (new_size - len(list))
